Refactor after updating from template
