+++ src/Mafin.Configuration/Providers/EnvironmentDependent/EnvironmentDependentConfigurationBuilderExtensions.cs
@@ -7,9 +7,9 @@ using Microsoft.Extensions.FileProviders;
 namespace Mafin.Configuration.Providers.EnvironmentDependent;
 
 /// <summary>
-/// Extension methods for <see cref="IConfigurationBuilder"/> to add environment dependent configuration.
+/// Provides extensions to <see cref="IConfigurationBuilder"/> for adding environment dependent configuration.
 /// </summary>
-public static class EnvironmentDependentConfigurationBuilderExtension
+public static class EnvironmentDependentConfigurationBuilderExtensions
 {
     /// <summary>
     /// Adds a environment dependent configuration source to <paramref name="builder"/>.
@@ -22,39 +22,39 @@ public static class EnvironmentDependentConfigurationBuilderExtension
             .GetCustomAttribute<AssemblyConfigurationAttribute>()?
             .Configuration!;
 
-        return AddEnvironmentDependentConfiguration(builder, EnvironmentConfigurationSplitPrinciple.ByFileName, environment, Environment.CurrentDirectory, optional: false, reloadOnChange: false, SearchOption.AllDirectories, provider: null);
+        return AddEnvironmentDependentConfiguration(builder, EnvironmentConfigurationSplit.ByFileName, environment, Environment.CurrentDirectory, optional: false, reloadOnChange: false, SearchOption.AllDirectories, provider: null);
     }
 
     /// <summary>
     /// Adds a environment dependent configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-    /// <param name="principle">The <see cref="EnvironmentConfigurationSplitPrinciple"/>.</param>
+    /// <param name="split">The <see cref="EnvironmentConfigurationSplit"/>.</param>
     /// <param name="baseDirectory">The <see cref="string"/> with base directory.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
     public static IConfigurationBuilder AddEnvironmentDependentConfiguration(
         this IConfigurationBuilder builder,
-        EnvironmentConfigurationSplitPrinciple principle,
+        EnvironmentConfigurationSplit split,
         string baseDirectory)
     {
         var environment = Assembly.GetCallingAssembly()?
             .GetCustomAttribute<AssemblyConfigurationAttribute>()?
             .Configuration!;
 
-        return AddEnvironmentDependentConfiguration(builder, principle, environment, baseDirectory, optional: false, reloadOnChange: false, SearchOption.AllDirectories, provider: null);
+        return AddEnvironmentDependentConfiguration(builder, split, environment, baseDirectory, optional: false, reloadOnChange: false, SearchOption.AllDirectories, provider: null);
     }
 
     /// <summary>
     /// Adds a environment dependent configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-    /// <param name="principle">The <see cref="EnvironmentConfigurationSplitPrinciple"/>.</param>
+    /// <param name="split">The <see cref="EnvironmentConfigurationSplit"/>.</param>
     /// <param name="baseDirectory">The <see cref="string"/> with base directory.</param>
     /// <param name="optional">Whether the file is optional.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
     public static IConfigurationBuilder AddEnvironmentDependentConfiguration(
         this IConfigurationBuilder builder,
-        EnvironmentConfigurationSplitPrinciple principle,
+        EnvironmentConfigurationSplit split,
         string baseDirectory,
         bool optional)
     {
@@ -62,21 +62,21 @@ public static class EnvironmentDependentConfigurationBuilderExtension
             .GetCustomAttribute<AssemblyConfigurationAttribute>()?
             .Configuration!;
 
-        return AddEnvironmentDependentConfiguration(builder, principle, environment, baseDirectory, optional, reloadOnChange: false, SearchOption.AllDirectories, provider: null);
+        return AddEnvironmentDependentConfiguration(builder, split, environment, baseDirectory, optional, reloadOnChange: false, SearchOption.AllDirectories, provider: null);
     }
 
     /// <summary>
     /// Adds a environment dependent configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-    /// <param name="principle">The <see cref="EnvironmentConfigurationSplitPrinciple"/>.</param>
+    /// <param name="split">The <see cref="EnvironmentConfigurationSplit"/>.</param>
     /// <param name="baseDirectory">The <see cref="string"/> with base directory.</param>
     /// <param name="optional">Whether the file is optional.</param>
     /// <param name="reloadOnChange">Whether the configuration should be reloaded if the file changes.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
     public static IConfigurationBuilder AddEnvironmentDependentConfiguration(
         this IConfigurationBuilder builder,
-        EnvironmentConfigurationSplitPrinciple principle,
+        EnvironmentConfigurationSplit split,
         string baseDirectory,
         bool optional,
         bool reloadOnChange)
@@ -85,14 +85,14 @@ public static class EnvironmentDependentConfigurationBuilderExtension
             .GetCustomAttribute<AssemblyConfigurationAttribute>()?
             .Configuration!;
 
-        return AddEnvironmentDependentConfiguration(builder, principle, environment, baseDirectory, optional, reloadOnChange, SearchOption.AllDirectories, provider: null);
+        return AddEnvironmentDependentConfiguration(builder, split, environment, baseDirectory, optional, reloadOnChange, SearchOption.AllDirectories, provider: null);
     }
 
     /// <summary>
     /// Adds a environment dependent configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-    /// <param name="principle">The <see cref="EnvironmentConfigurationSplitPrinciple"/>.</param>
+    /// <param name="split">The <see cref="EnvironmentConfigurationSplit"/>.</param>
     /// <param name="environment">The <see cref="string"/> with environment.</param>
     /// <param name="baseDirectory">The <see cref="string"/> with base directory.</param>
     /// <param name="optional">Whether the file is optional.</param>
@@ -102,7 +102,7 @@ public static class EnvironmentDependentConfigurationBuilderExtension
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
     public static IConfigurationBuilder AddEnvironmentDependentConfiguration(
         this IConfigurationBuilder builder,
-        EnvironmentConfigurationSplitPrinciple principle,
+        EnvironmentConfigurationSplit split,
         string environment,
         string baseDirectory,
         bool optional,
@@ -110,16 +110,16 @@ public static class EnvironmentDependentConfigurationBuilderExtension
         SearchOption searchOption,
         IFileProvider? provider)
     {
-        switch (principle)
+        switch (split)
         {
-            case EnvironmentConfigurationSplitPrinciple.ByFolder:
+            case EnvironmentConfigurationSplit.ByFolder:
                 builder.AddDirectoryProbing(new[] { Path.Combine(environment, "*.*") }, baseDirectory, formatSourceMap: null, searchOption, provider, optional, reloadOnChange);
                 break;
-            case EnvironmentConfigurationSplitPrinciple.ByFileName:
+            case EnvironmentConfigurationSplit.ByFileName:
                 builder.AddDirectoryProbing(new[] { $"*.{environment}.*" }, baseDirectory, formatSourceMap: null, searchOption, provider, optional, reloadOnChange);
                 break;
             default:
-                throw new ArgumentOutOfRangeException(nameof(principle), principle, "Unsupported value");
+                throw new ArgumentOutOfRangeException(nameof(split), split, "Unsupported value");
         }
 
         return builder;
