+++ src/Mafin.Configuration/Providers/Toml/TomlConfigurationBuilderExtensions.cs
@@ -1,55 +1,42 @@
+using Mafin.Configuration.Parsers;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Configuration.Memory;
 using Microsoft.Extensions.FileProviders;
 
-namespace Mafin.Configuration.Providers.SettingsFile;
+namespace Mafin.Configuration.Providers.Toml;
 
 /// <summary>
-/// Extension methods for <see cref="IConfigurationBuilder"/> to add <see cref="SettingsFileConfigurationSource"/>.
+/// Provides extensions to <see cref="IConfigurationBuilder"/> for adding <see cref="TomlConfigurationSource"/>.
 /// </summary>
-public static class SettingsFileConfigurationBuilderExtension
+public static class TomlConfigurationBuilderExtensions
 {
-    private const string ConfigurationFileName = "Mafin.Configuration.json";
-
-    private static string DefaultSettingsFile =>
-        Path.Combine(Environment.CurrentDirectory, ConfigurationFileName);
-
     /// <summary>
-    /// Adds configuration source parameters from module settings file to <paramref name="builder"/>.
-    /// </summary>
-    /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
-    /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-    public static IConfigurationBuilder LoadSettingsFile(this IConfigurationBuilder builder)
-    {
-        return LoadSettingsFile(builder, DefaultSettingsFile);
-    }
-
-    /// <summary>
-    /// Adds configuration source parameters from module settings file to <paramref name="builder"/>.
+    /// Adds a TOML configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
     /// <param name="path">Path relative to the base path stored in
     /// <see cref="IConfigurationBuilder.Properties"/> of <paramref name="builder"/>.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-    public static IConfigurationBuilder LoadSettingsFile(this IConfigurationBuilder builder, string path)
+    public static IConfigurationBuilder AddTomlFile(this IConfigurationBuilder builder, string path)
     {
-        return LoadSettingsFile(builder, provider: null, path: path, optional: false, reloadOnChange: false);
+        return AddTomlFile(builder, provider: null, path: path, optional: false, reloadOnChange: false);
     }
 
     /// <summary>
-    /// Adds configuration source parameters from module settings file to <paramref name="builder"/>.
+    /// Adds a TOML configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
     /// <param name="path">Path relative to the base path stored in
     /// <see cref="IConfigurationBuilder.Properties"/> of <paramref name="builder"/>.</param>
     /// <param name="optional">Whether the file is optional.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-    public static IConfigurationBuilder LoadSettingsFile(this IConfigurationBuilder builder, string path, bool optional)
+    public static IConfigurationBuilder AddTomlFile(this IConfigurationBuilder builder, string path, bool optional)
     {
-        return LoadSettingsFile(builder, provider: null, path: path, optional: optional, reloadOnChange: false);
+        return AddTomlFile(builder, provider: null, path: path, optional: optional, reloadOnChange: false);
     }
 
     /// <summary>
-    /// Adds configuration source parameters from module settings file to <paramref name="builder"/>.
+    /// Adds a TOML configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
     /// <param name="path">Path relative to the base path stored in
@@ -57,13 +44,13 @@ public static class SettingsFileConfigurationBuilderExtension
     /// <param name="optional">Whether the file is optional.</param>
     /// <param name="reloadOnChange">Whether the configuration should be reloaded if the file changes.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-    public static IConfigurationBuilder LoadSettingsFile(this IConfigurationBuilder builder, string path, bool optional, bool reloadOnChange)
+    public static IConfigurationBuilder AddTomlFile(this IConfigurationBuilder builder, string path, bool optional, bool reloadOnChange)
     {
-        return LoadSettingsFile(builder, provider: null, path: path, optional: optional, reloadOnChange: reloadOnChange);
+        return AddTomlFile(builder, provider: null, path: path, optional: optional, reloadOnChange: reloadOnChange);
     }
 
     /// <summary>
-    /// Adds configuration source parameters from module settings file to <paramref name="builder"/>.
+    /// Adds a TOML configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
     /// <param name="provider">The <see cref="IFileProvider"/> to use to access the file.</param>
@@ -72,8 +59,7 @@ public static class SettingsFileConfigurationBuilderExtension
     /// <param name="optional">Whether the file is optional.</param>
     /// <param name="reloadOnChange">Whether the configuration should be reloaded if the file changes.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-    public static IConfigurationBuilder LoadSettingsFile(this IConfigurationBuilder builder, IFileProvider? provider, string path, bool optional, bool reloadOnChange)
-#pragma warning disable PH2071 // Avoid Duplicate Code
+    public static IConfigurationBuilder AddTomlFile(this IConfigurationBuilder builder, IFileProvider? provider, string path, bool optional, bool reloadOnChange)
     {
         if (builder == null)
         {
@@ -85,23 +71,51 @@ public static class SettingsFileConfigurationBuilderExtension
             throw new ArgumentException($"'{nameof(path)}' cannot be null or empty.", nameof(path));
         }
 
-        return builder.LoadSettingsFile(s =>
+        return builder.AddTomlFile(s =>
         {
             s.FileProvider = provider;
             s.Path = path;
             s.Optional = optional;
             s.ReloadOnChange = reloadOnChange;
             s.ResolveFileProvider();
-#pragma warning restore PH2071 // TODO: Refactor code to remove duplication
         });
     }
 
     /// <summary>
-    /// Adds configuration source parameters from module settings file to <paramref name="builder"/>.
+    /// Adds a TOML configuration source to <paramref name="builder"/>.
     /// </summary>
     /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
     /// <param name="configureSource">Configures the source.</param>
     /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
-    public static IConfigurationBuilder LoadSettingsFile(this IConfigurationBuilder builder, Action<SettingsFileConfigurationSource> configureSource)
+    public static IConfigurationBuilder AddTomlFile(this IConfigurationBuilder builder, Action<TomlConfigurationSource> configureSource)
         => builder.Add(configureSource);
+
+    /// <summary>
+    /// Adds a TOML configuration source to <paramref name="builder"/>.
+    /// </summary>
+    /// <param name="builder">The <see cref="IConfigurationBuilder"/> to add to.</param>
+    /// <param name="stream">The <see cref="Stream"/> to read the json configuration data from.</param>
+    /// <returns>The <see cref="IConfigurationBuilder"/>.</returns>
+    public static IConfigurationBuilder AddTomlStream(this IConfigurationBuilder builder, Stream stream)
+    {
+        if (builder == null)
+        {
+            throw new ArgumentNullException(nameof(builder), $"'{nameof(builder)}' cannot be null");
+        }
+
+        IDictionary<string, string?> data;
+
+        try
+        {
+            using var reader = new StreamReader(stream);
+            data = new TomlParser().Parse(reader.ReadToEnd());
+        }
+        catch (Exception e)
+        {
+            throw new FormatException("Unable to parse configuration file in TOML format.", e);
+        }
+
+        var source = new MemoryConfigurationSource() { InitialData = data };
+        return builder.Add(source);
+    }
 }
