--- tests/Mafin.Configuration.Tests.Unit/UnitTest1.cs
@@ -1,10 +1,0 @@
-namespace Mafin.Configuration.Tests.Unit;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-        Assert.True(true);
-    }
-}
